HTML pages
This is the pages that were given to us via link in canvas.
this should be everything, i added an index.html redirect page. I plan
on adding links to each individual page on the index.html and removing
the redirect so each page can have a quick jump to.