modified index.html with just links to each page
I took a moment to create a list of all of the pages so they can be
viewed as currently coded. there is leftover code in these that i choose
to not delete (search results and reviews to be specific) also while
saving my self from typing i noticed that screen1.php was missing the
opening html tag so i added it.