user login screen
deleted a file that was not needed and attempted to work on the redirect
for the user login screen. for some reason the redirect from the post is
not working. i tried several things but i could not get it to work. i
know the SQL is good and the logic is good for the if statement but i
can't get it to function.